update witherbone
the base bone option can be used as either main or as rod, the witherbone has the exact same stats as the vanilla bone for rod other than the tier, but cant be a main part, which means that the witherbone even though harder to acquire is practically a lesser component that can be used in more powerful tools.
this gives a minor (.05%) buff to the part. I would also potentially recommend adding a trait for giving wither effect for a small period of time.